+++ CHANGELOG.md
@@ -9,6 +9,12 @@ All notable changes to this project will be documented in this file.  The format
 [comment]: <> (Fixed:      any bug fixes)
 [comment]: <> (Security:   in case of vulnerabilities)
 
+## [1.1.2]
+
+### Added
+* New RPC method `GetEraSummary()`.
+* New global state key `era-summary-0000...0000`.
+
 ## [1.1.1]
 
 ### Fixed
@@ -33,6 +39,7 @@ All notable changes to this project will be documented in this file.  The format
 ### Added
 * Initial release of Casper .NET SDK.
 
+[1.1.2]: https://github.com/make-software/casper-net-sdk/releases/tag/v1.1.2
 [1.1.1]: https://github.com/make-software/casper-net-sdk/releases/tag/v1.1.1
 [1.1.0]: https://github.com/make-software/casper-net-sdk/releases/tag/v1.1.0
 [1.0.0]: https://github.com/make-software/casper-net-sdk/releases/tag/v1.0.0